third_party/WebKit/LayoutTests/external/wpt/html/browsers/browsing-the-web/history-traversal/browsing_context_name_cross_origin_2-expected.txt

This is a testharness.js-based test.
FAIL Restoring window.name on cross-origin history traversal assert_equals: After navigation expected "test3" but got "test"
Harness: the test ran to completion.
